clk: qcom: mmcc-msm8998: Disable safe config on maxi and mdp clocks
It is not necessary to park these clocks to XO when shutting them
down as we expect the right PLL to be up when we resume them.

Also, it's not necessary to swtich to XO before switching to
the new PLL input, because temporary clock unstability on these
RCGs is not gonna lock up anything.