regulator: rpm-smd: Revert LDOA regulator changes
Our device tree reflects overrides such as pin ctrl mode and
level parameters of which have been disabled with the introduction
of set_load. This revision may be causing various LDOA regulators
to malfunction, even if we re-enable the overrides in the parameter
table as this revision is likely interfering with these overrides.

As observed, this causes instability in situations wherein mobile
data is heavily used during low signal instances. Leading to RIL
completely collapsing then recovering, and so on.

The k4.4 revision provides better stability in such cases at least
in practical tests.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>