clk: qcom: mmcc-msm8998: Enable safe config on all source clocks
This is needed to take into account the HW limitation wherein
MM clocks are reparented towards XO after power collapse. What
this does is that it allows the MM drivers to configure the rate
properly prior to enabling the RCG and actually have the clocks
be enabled on the proper parent clock determined according to the
set_rate performed prior.

The HW limitation was stated in the following commit in downstream:
https://github.com/ederevx/x_kernel_oneplus_msm8998/commit/52168e4c325387c27ace51956752462f0898b2b8

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>