clk: qcom: gcc-msm8998: Control the HMSS AHB clock based on mainline
Let's make sure that this clock never turns off by registering it
and voting on it explicitly. This is done in mainline, and this
is what actually makes Kholk's regmap patching at init valid since
setting sleep bit ena is necessary to allow HW to gate the clock
if Linux registers the clock and configures it to be always on.

This will help ensure the platform doesn't die or crash just in
case RPM misbehaves.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>