arm64: DT: msm8998-regulator: Set PM8998 LDO 26 as always-on for msm8998
This powers the UFS device's ref_clk pad power wherein it must be
kept alive for as long as the device is running, until the device
gets reset. sdm845 has configured this to be always-on for this
reason so also do it for our platform but ours would be on the L26
regulator (vddp-ref-clk-supply).

Reference:
https://github.com/LineageOS/android_kernel_xiaomi_sdm845/commit/413af2c2e9235b1376a5e739b7d029ec6c3ca20d

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>