arm64: DT: pm8998-rpm-regulator: Correct HPM min loads for LDO 2 and 26
sdm845 had actually swapped the roles of LDO 2 and 26 according
to the device tree, at least with the UFS. Furthermore, LDO 2
is actually also providing power to multimedia consumers which are
beyond the APSS processor - and thus we should ensure to be at HPM
once any load value has been set. Hence, let's increase the HPM value
on LDO 26 instead as it is only connected to the platform's UFS device.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>