power: qcom: op_cg_uovp: Check usb_icl_votable after voting to ensure it applies
The vote function's return code only accounts for whether the vote
has been taken into account by the USB driver but does not indicate
that the SW is using the ICHG_UA value that we have voted - likely
due to HW limitations. Make op_cg_current_set() aware of this and
have it check the effective result after voting to ensure the
ichg_ua value voted matches it.

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>