drivers: thermal: lmh_dcvs: Disable min_freq_reg platform mitigation for legacy platforms
This does not exist in downstream [1] and it was introduced for sdm845
[2] which could have been added to msm8998 if it was supported since
floor_limit call function already exists in downstream and can be
added - but it was not added in msm8998 nor sdm660.

Disable floor limit thermal configuration to avoid touching regs
in legacy platforms that we're not supposed to touch. This will
switch back floor mitigation to CPUFREQ.

References:
[1]: https://github.com/LineageOS/android_kernel_oneplus_msm8998/blob/lineage-22.1/drivers/thermal/msm_lmh_dcvs.c
[2]: https://github.com/LineageOS/android_kernel_xiaomi_sdm845/commit/a5911a383fac31241a469ff1e58eaeff04d02499

Signed-off-by: Edrick Sinsuan <evcsinsuan@gmail.com>